+++ CHANGELOG.md
@@ -1,8 +1,21 @@
 ## Unreleased
 
+BUG FIXES:
+* Fix decoding fields for StorageCallVShardError (MasterUUID, ReplicasetUUID).
+
+CHANGES:
+* Add comment why and how we handle "NON_MASTER" vshard error.
+* Don't support 'type Error struct' anymore.
+* Handle vshard error the same way as lua vshard router (resolve issue #77).
+
 FEATURES:
 
 * Add pause between requests in buckets discovering. Configured by config DiscoveryWorkStep, default is 10ms.
+* Add ReplicaUUID to the StorageCallVShardError struct.
+
+REFACTOR:
+
+* Use constants for vshard error names and codes.
 
 ## v1.2.0